add pagination for lots and comments

+++ src/main/java/ru/koChyan/Auction/controller/CommentController.java
@@ -2,6 +2,10 @@ package ru.koChyan.Auction.controller;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.web.PageableDefault;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -13,6 +17,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import ru.koChyan.Auction.controller.util.ControllerUtils;
 import ru.koChyan.Auction.domain.Comment;
 import ru.koChyan.Auction.domain.Lot;
+import ru.koChyan.Auction.domain.Status;
 import ru.koChyan.Auction.domain.User;
 import ru.koChyan.Auction.service.CommentService;
 
@@ -31,12 +36,17 @@ public class CommentController {
     @GetMapping()
     public String commentList(
             @PathVariable Lot lot,
-            Model model
+            Model model,
+            @PageableDefault(sort = {"id"}, direction = Sort.Direction.DESC) Pageable pageable
     ) {
-
-        model.addAttribute("lot", lot);
-        model.addAttribute("comments", commentService.getAllByLotId(lot.getId()));
-        return "comment/lotCommentList";
+        if (lot.getStatus().equals(Status.ACTIVE.name())) { // если лот все еще активен
+            model.addAttribute("lot", lot);
+            model.addAttribute("url", "/lot/" + lot.getId() + "/comment");
+            model.addAttribute("page", commentService.getAllByLotId(lot.getId(), pageable));
+            return "comment/lotCommentList";
+        } else {
+            return "redirect:/lot";
+        }
     }
 
     @PostMapping()
@@ -45,30 +55,41 @@ public class CommentController {
             @PathVariable Lot lot,
             Model model,
             @Valid Comment comment,
-            BindingResult bindingResult
+            BindingResult bindingResult,
+            @PageableDefault(sort = {"id"}, direction = Sort.Direction.DESC) Pageable pageable
+
     ) {
         if (bindingResult.hasErrors()) {
             Map<String, List<String>> errors = ControllerUtils.getErrors(bindingResult);
 
             model.mergeAttributes(errors);
             model.addAttribute("lot", lot);
-            model.addAttribute("comments", commentService.getAllByLotId(lot.getId()));
+            model.addAttribute("url", "/lot/" + lot.getId() + "/comment");
+            model.addAttribute("page", commentService.getAllByLotId(lot.getId(), pageable));
             return "comment/lotCommentList";
         } else {
+            if (lot.getStatus().equals(Status.ACTIVE.name())) { //если лот все еще активен
 
-            commentService.addComment(comment, lot, user);
-            return "redirect:/lot/" + lot.getId() + "/comment";
+                commentService.addComment(comment, lot, user);
+                return "redirect:/lot/" + lot.getId() + "/comment";
+            } else {
+                return "redirect:/lot";
+            }
         }
     }
 
+    @PreAuthorize("hasAuthority('ADMIN')") // требуем права админа для удаления комментария
     @PostMapping("/{comment}/delete")
     public String deleteComment(
             @PathVariable Lot lot,
             @PathVariable Comment comment
     ) {
-
-        commentService.remove(comment);
-        return "redirect:/lot/" + lot.getId() + "/comment";
+        if (lot.getStatus().equals(Status.ACTIVE.name())) { //если лот все еще активен
+            commentService.remove(comment);
+            return "redirect:/lot/" + lot.getId() + "/comment";
+        } else {
+            return "redirect:/lot";
+        }
     }
 
 }
